Allow custom mappers
In the README it states:

> Nothing is stopping you from decorating your structs using registered
> mappers and custom decorator models:

However this doesn't appear to work as advertised. If you have a
custom mapper and send to a non-virtus model it will fail. If you have
custom mapping inside your mapper eg:

```ruby
  attribute :jsonb_column do |td|
    JSON.parse(td, symbolize_names: true)
  end
```

Then it fails in transproc. In both situations the process is:

```
Call method on repository
  |
  --> Calls relation view to retrieve data
      |
      --> Maps that using ROM Repository's mapper
          |
          --> ROM::Struct
              |
              --> calls `as(:custom_mapper)`
                  |
                  --> This mapper expects a hash, not a struct blows up
```

It's not an ideal solution, but delegating method missing to the hash
fixes all these problems in the short term.